requirements-dev.txt
amqp==2.4.1
asgiref==3.2.3
billiard==3.5.0.5
celery==4.2.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
dataclasses==0.6
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
django-celery-beat==1.5.0
django-cors-headers==3.0.1
django-extensions==2.1.4
django-rest-framework==0.1.0
django-serializable-model==0.0.3
djangorestframework==3.9.0
graphviz==0.10.1
gunicorn==19.9.0
kombu==4.2.2.post1
llvmlite==0.31.0
Markdown==3.0.1
numba==0.48.0
numpy==1.18.2
opencv-contrib-python==4.2.0.32
opencv-python==3.4.3.18
Pillow==5.4.1
psycopg2==2.8.2
pycparser==2.19
pydot==1.4.1
pydotplus==2.0.2
pyparsing==2.3.0
PyQt5==5.12
PyQt5-sip==4.19.14
python-crontab==2.4.0
python-dateutil==2.8.1
python-decouple==3.1
python-qt5==0.1.10
pytz==2018.7
selenium==3.141.0
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
typing==3.6.6
urllib3==1.25.8
vine==1.2.0
whitenoise==4.1.2
